A4/backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.4.4
